Moved the styles to a separate folder with styles for the find-projects page
