+++ .env
@@ -1,0 +1,4 @@
+NEXT_PUBLIC_BACKEND_API_URL=http://hubdev.svyat404.com
+
+NEXTAUTH_SECRET=somesalt
+NEXTAUTH_JWT_AGE=10000
