+++ next.config.mjs
@@ -5,4 +5,6 @@ const nextConfig = {
   },
 };
 
+const NEXTAUTH_URL='http://localhost:3001'
+
 export default nextConfig;
