+++ package.json
@@ -19,7 +19,7 @@
     "@mui/lab": "^5.0.0-alpha.147",
     "@mui/material": "^5.14.11",
     "next": "13.4.3",
-    "next-auth": "^4.22.1",
+    "next-auth": "^4.24.5",
     "normalize.css": "^8.0.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
