+++ src/components/layout/breadcrumbs/Breadcrams.tsx
@@ -1,8 +1,10 @@
-'use client'
-import { Options } from '../Types/Types';
+'use client';
+
+import React from 'react';
 import { Box, Breadcrumbs, Typography } from '@mui/material';
 import Link from 'next/link';
 import { useSearchParams } from 'next/navigation';
+import { Options } from '../../../Types/Types';
 
 type Props = {
   options: Options
@@ -10,16 +12,16 @@ type Props = {
 
 const Breadcrams: React.FC<Props> = (props) => {
   const searchParams = useSearchParams();
-  const projectType = searchParams.get("value");
+  const projectType = searchParams.get('value');
 
   return (
     <Box sx={breadcrumbs.box}>
-      <Breadcrumbs aria-label="breadcrumbs" separator="·">
+      <Breadcrumbs aria-label='breadcrumbs' separator='·'>
         <Link style={breadcrumbs.link} href={{ pathname: '/find-project' }}>Найти проект</Link>
         {projectType ? <Typography sx={breadcrumbs.active}>{props.options[projectType]}</Typography> : ''}
       </Breadcrumbs>
     </Box>
-  )
+  );
 };
 export default Breadcrams;
 
@@ -38,5 +40,4 @@ const breadcrumbs = {
   active: {
     color: ' #647380',
   },
-}
-
+};
